Introduce react-router-dom, refactor landing page to a separate component, make route to the app root forwarding to it
