+++ frontend/package.json
@@ -12,6 +12,7 @@
     "axios": "^1.3.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.9.0",
     "sass": "^1.59.3"
   },
   "devDependencies": {
