+++ frontend/src/index.scss
@@ -1,3 +1,3 @@
-@use "src/sass/variables";
+@use "sass/variables";
 @use "sass/global";
 @use "sass/components";
