+++ frontend/src/main.tsx
@@ -1,10 +1,13 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
+import {BrowserRouter as Router} from 'react-router-dom'
 import App from './components/App/App'
 import './index.scss'
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
+	<React.StrictMode>
+		<Router>
+			<App/>
+		</Router>
+	</React.StrictMode>,
 )
