+++ frontend/src/sass/components.scss
@@ -1,1 +1,2 @@
 @use "../components/App/App";
+@use "../components/LandingPage/LandingPage";
